+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   Added E8010 to avoid using `os.popen()` as it internally uses `subprocess.Popen` with `shell=True`
+-   Added E8011 to avoid using `shlex.quote()` on non-POSIX platforms.
 
 ## [1.1.0] - 2021-07-02